对输入的时间进行判断

+++ com/xgz/now.py
@@ -11,7 +11,7 @@ def update_time(time):
     :return:
     """
     try:
-        if type(time) == int:
+        if type(time) == int and int(time) >= 2:
             revise_yaml(f'time: {time}', yam['Label']['time'])
             os.system("kill -9 $(netstat -nlp | grep :5000 | awk '{print $7}' | awk -F'/' '{ print $1 }')")
     except Exception as e:
